==> R/02_clean.qmd <==
---
title: "02_clean"
format: html
embed-resources: true
---

## Libraries

```{r}
#| message: false

library("here")
library("tidyverse")
```

## Expression data

For the analysis only the probe ID and signal is needed. The p-value is used to choose which probes to work with. The file identifier is needed to combine the expression data and the metadata. Two people did not complete the trial and as a result there is no metadata for them so their expression data is removed.

```{r}
#| message: false

data_expression <- read_tsv(here("data", "01_expression_dat_load.tsv.gz")) |> 
  select(file, PROBE_ID, Signal, Pval) |> 
  filter(file != "GSM1008315" & file != "GSM1008340")
```

For the analysis only probes with an average p-value under 0.05 are used

```{r}
data_expression <- data_expression |>
  group_by(PROBE_ID) |> 
  nest() |> 
  mutate(Avg_Pval = map_dbl(.x = data, 
                            .f = ~mean(pull(.x, Pval)))) |> 
  unnest(cols = c(data)) |> 
  ungroup() |>
  filter(Avg_Pval < 0.05) |> 
  select(-Avg_Pval)
```

```{r}
write_tsv(data_expression, here("data", "02_expression_dat_clean.tsv.gz"))
```

## Metadata

```{r}
#| message: false

metadata <- read_tsv(here("data", "01_meta_dat_load.tsv.gz"))

metadata <- metadata |> 
  select(-starts_with("Season"),
        -starts_with("VacType"),
        -Prior_vac,
        -VACCOUNT,
        -FLU_EVR,
        -FLU_HOSP)

write_tsv(metadata, here("data", "02_meta_dat_clean.tsv"))
```
