==> R/03_augment.qmd <==
---
title: "03_augment"
format: html
embed-resources: true
---

```{r}
#| message: false

library("tidyverse")
library("here")
```

Load the data from the final files.

```{r}
#| message: false
metadata <- read_tsv(here("data", "02_meta_dat_clean.tsv"))
analysis_data <- read_tsv(here("data", "02_expression_dat_clean.tsv.gz"))
```

In order to create the final dataset, we need to pivot wider the data with mRNA expression using as row names the sample IDs, column names will be the Probe IDs and the values are the signals for each Probe.

```{r}
analysis_data <- analysis_data |> 
  pivot_wider(id_cols = file,         
              names_from = PROBE_ID,  
              values_from = Signal) 
```

We, also, need to connect the two datasets by the IDs.

```{r}
mapping_key <- analysis_data |> 
  select(file) |>
  distinct()

metadata <- metadata |> 
  mutate(mapping_key) |> 
  select(-xID)
```

Now we can connect the two datasets using as key to join the mapping_key created.

```{r}
analysis_data <- metadata |> 
  full_join(analysis_data, by = "file") |> 
  relocate("file") 

analysis_data <- analysis_data |>
  rename(EBV_pos = "EBV (1 = pos)",
         CMV_pos = "CMV (1= pos)")

analysis_data <- analysis_data |>
  mutate(
    Age_Group = case_when(
      Age >= 20 & Age <= 30 ~ "Young",
      Age >= 61 ~ "Older"),
    Cytomegalovirus = case_when(
      CMV_pos == 1 ~ "Positive",
      CMV_pos == 0 ~ "Negative"),
    EpsteinBarrvirus = case_when(
      EBV_pos == 1 ~ "Positive",
      EBV_pos == 0 ~ "Negative")) |> 
  relocate(Age_Group, .after = Age) |> 
  relocate(Cytomegalovirus, .after = BMI) |> 
  relocate(EpsteinBarrvirus, .after = BMI ) |> 
  select(-`CMV_pos`, -EBV_pos) 

```

The code checks whether each person had a 4-fold antibody increase for the three flu strains, counts how many strains they responded to, and labels them as poor or good responders. It then reorganizes these new columns and summarizes response rates by age group and number of strains for plotting.

```{r}
analysis_data <- analysis_data |>
  mutate(
    H1N1_fold = (H1N1v3 / H1N1v1) >= 4,
    B_fold    = (Bv3 / Bv1) >= 4,
    H3N2_fold = (H3N2v3 / H3N2v1) >= 4,
    Strains = H1N1_fold + B_fold + H3N2_fold,
    Response_group = if_else(Strains <= 1, "Poor responders", "Good responders")
  )

analysis_data <- analysis_data |>
  relocate(H1N1_fold, B_fold, H3N2_fold, Strains, Response_group,
           .after = Cytomegalovirus)

```

```{r}
write_tsv(analysis_data, here("data", "03_expression_dat_aug.tsv.gz"))

rm(list = ls())
```
