==> R/04_describe.qmd <==
---
title: "04_describe"
format: html
embed-resources: true
---

Loading the libraries we will need

```{r}
#| message: false
library("table1")
library("tidyverse")
library("here")
source("99_proj_func.R")

Sys.setenv("VROOM_CONNECTION_SIZE" = 131072 * 2)
```

Loading the dataset we are going to analyse.

```{r}
#| message: false
analysis_data <- read_tsv(here("data", "03_expression_dat_aug.tsv.gz"))

```

To describe the dataset, we are making two tables. The first table shows the average age and BMI of the two groups, Old and young, as well as the range of each measure in both groups. The second table shows the distribution of each group in the three variables: Gender, Cytomegalovirus and Epstein-Barr virus (positive/negative).

```{r}
numeric_summary <- analysis_data |>
  group_by(Age_Group) |>
  summarize(Age_Median_Range = str_c(
      round(median(Age, na.rm = TRUE), 1), 
      " (", 
      min(Age, na.rm = TRUE), 
      "–", 
      max(Age, na.rm = TRUE), 
      ")"),
    BMI_Median_Range = str_c(
      round(median(BMI, na.rm = TRUE), 1), 
      " (", 
      min(BMI, na.rm = TRUE), 
      "–", 
      max(BMI, na.rm = TRUE), 
      ")"),
    .groups = 'drop')

gender_table <- get_categorical_summary(analysis_data, Gender) |> 
  mutate(Variable = "Gender")
cmv_table <- get_categorical_summary(analysis_data, Cytomegalovirus) |> 
  mutate(Variable = "Cytomegalovirus")
ebv_table <- get_categorical_summary(analysis_data, EpsteinBarrvirus) |> 
  mutate(Variable = "EpsteinBarrvirus")

categorical_summary <- bind_rows(gender_table, cmv_table, ebv_table)

final_categorical_table <- categorical_summary |>
  pivot_wider(names_from = Age_Group,
              values_from = Value,
              id_cols = c(Variable, Characteristic),
              names_sort = TRUE)

print(numeric_summary)
print(final_categorical_table)
```
