==> R/05_analysis1.qmd <==
---
title: "05_analysis1"
format: html
embed-resources: true
---

### Libraries

```{r}
#| message: false

library("tidyverse")
library("here")
library("broom")
library("patchwork")
library("ggridges")

Sys.setenv("VROOM_CONNECTION_SIZE" = 131072 * 2)
```

### PCA

A PCA on just the signals from the probes is done to see if it can show a connection between the signals pre vaccine and the response to the vaccine. The PCA is done with inspiration from ["PCA tidyverse style" by Claus O. Wilke](https://clauswilke.com/blog/2020/09/07/pca-tidyverse-style/)

```{r}
#| message: false
analysis_data <- read_tsv(here("data", "03_expression_dat_aug.tsv.gz"))
```

```{r}
pca <- analysis_data |> 
  select(starts_with("ILMN_")) |> 
  scale() |> 
  prcomp()
```

#### Variance explained by principal components

```{r}
pca_var <- pca |> 
  tidy("eigenvalues")

p1 <- pca_var |> 
  ggplot(aes(x = PC, 
             y = percent,
             fill = "#ED6F58",
             color = "#ED6F58")) +
  geom_col() +
  labs(title = "Individual",
       subtitle = "All PCs",
       y = "Explained Variance") +
  theme_minimal() +
  theme(legend.position = "none")

p2 <- pca_var |> 
  filter(PC < 6) |> 
  ggplot(aes(x = PC, 
             y = percent,
             fill = "#ED6F58",
             color = "#ED6F58")) +
  geom_col() +
  labs(title = "Individual",
       subtitle = "5 first PCs",
       y = "Explained Variance") +
  theme_minimal() +
  theme(legend.position = "none")

p3 <- pca_var |> 
  ggplot(aes(x = PC, 
             y = cumulative)) +
  geom_line() +
  geom_point(color = "#ED6F58") +
  labs(title = "Cumulative",
       y = "Explained Variance") +
  scale_y_continuous(breaks = seq(from = 0.4,
                                  to = 1,
                                  by = 0.1)) +
  theme_minimal() +
  theme(legend.position = "none")

p1 / (p2 + p3)
```

```{r}
#| message: false

ggsave(filename = "05_variance_plot.png",
       path = here("results"))
```

The cumulative plot shows that the first 5 principal components explain over 70% of the variance. As a result they are examined to see if they can be used to separate good and poor responders.

#### Data in principal component coordinates

First the data in the coordinates for the 5 first principal components is split into good or poor response to find the 2 principal components that are most likely to separate them.

```{r}
#| message: false

pca |> 
  augment(analysis_data) |> 
  pivot_longer(cols = starts_with(".fitted"),
               names_to = "PC",
               values_to = "Coordinates") |> 
  mutate(PC = as.numeric(str_extract(PC,
                                     "[0-9]+"))) |> 
  filter(PC < 6) |> 
  mutate(PC = factor(PC)) |> 
  ggplot(aes(x = Coordinates, 
             y = PC, 
             fill = Response_group,
             color = Response_group)) +
  geom_density_ridges(alpha = 0.5,
                      scale = 1.2) +
  scale_fill_manual(values = c("#E69F00", "#0072B2")) +
  scale_color_manual(values = c("#E69F00", "#0072B2")) +
  theme_minimal() +
  theme(legend.position = "bottom",
        legend.title = element_blank())
```

```{r}
#| message: false

ggsave(filename = "05_1d_plot.png",
       path = here("results"))
```

The plot shows that none of the 5 first principal componets separate the good and poor responders well. Principal component 2 and 3 are chosen as they show most separation and maybe together can separate the type of response better.

```{r}
pca |> 
  augment(analysis_data) |> 
  rename_with(.cols = starts_with(".fitted"),
              .fn = ~str_extract(.x,
                                 "PC[0-9]+")) |> 
  ggplot(aes(x = PC2, 
             y = PC3, 
             color = Response_group)) +
  geom_point(size = 3) +
  scale_color_manual(values = c("#E69F00", "#0072B2")) +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.box.background = element_rect(color = "lightgrey"))
```

```{r}
#| message: false

ggsave(filename = "05_2d_plot.png",
       path = here("results"))
```

The plot shows no separation so the PCA does not show a connection between the signals pre vaccine and the response to it.
