==> R/06_analysis2.qmd <==
---
title: "06_analysis2"
format: html
embed-resources: true
---

To examine the connection between poor or good response to the vaccine and some of the metadata the figures in figure 2 in the paper are recreated

Plot A:

```{r}
#| message: false
library("tidyverse")
library("here")

Sys.setenv("VROOM_CONNECTION_SIZE" = 131072 * 2)
```

```{r}
#| message: false
analysis_data <- read_tsv(here("data", "03_expression_dat_aug.tsv.gz"))
```

```{r}
plot_data <- analysis_data |>
  group_by(Age_Group, Strains) |>
  summarise(n = n(), .groups = "drop") |>
  group_by(Age_Group) |>
  mutate(rate = 100 * n / sum(n)) |>
  ungroup() |>
  mutate(Strains = factor(Strains, levels = 0:3))
```

```{r}
ggplot(plot_data,
       aes(x = as.numeric(as.character(Strains)),
           y = rate,
           fill = Age_Group)) +
  geom_col(position = "dodge", colour = "black", width = 0.7) +
  scale_x_continuous(breaks = 0:3, labels = 0:3) +
  scale_fill_manual(values = c("Older" = "white", "Young" = "black")) +
  geom_vline(xintercept = 1.5, linetype = "dashed") +
  labs(x = "Number of strains",
       y = "Seroconversion rate (%)",
       fill = NULL) +
  theme_bw() +
  theme(panel.grid = element_blank())

```

If only 0 or 1 of the strains had a 4-fold antibody increase they were labeled poor responders (left of the dotted line). The good responders were those where 2 or 3 of the strains had a 4-fold antibody increase (right of the dotted line).

```{r}
#| message: false

ggsave(filename = "06_A_plot.png",
       path = here("results"))
```

Plot B:

```{r}
plot_b_data <- analysis_data |>
  count(Age_Group, Response_group) |>
  group_by(Age_Group) |>
  mutate(percent = 100 * n / sum(n)) |>
  ungroup() |>
  mutate(
    Age_Group = factor(Age_Group, levels = c("Young", "Older")),
    Response_group = factor(Response_group, levels = c("Poor responders", "Good responders"))  
  )

ggplot(plot_b_data, aes(y = Age_Group, x = percent, fill = Response_group)) +
  geom_col(position = "fill", colour = "black", width = 0.6) +
  scale_x_continuous(labels = scales::percent_format()) +
  scale_fill_manual(
    values = c("Poor responders"="white", "Good responders" = "black"),
    labels = c("PR", "GR")
  ) +
  labs(x = "Percent response", y = NULL, fill = NULL) +
  theme_bw() +
  theme(panel.grid = element_blank())
```

There is larger percent of the young age group that are good responders than in the older age group. The color of the plot makes it difficult to see that the white is also a group so it is improved with color.

```{r}
#| message: false

ggsave(filename = "06_B1_plot.png",
       path = here("results"))
```

Improved version of plot B:

```{r}
plot_b_data <- analysis_data |>
  count(Age_Group, Response_group) |>
  group_by(Age_Group) |>
  mutate(percent = 100 * n / sum(n)) |>
  ungroup() |>
  mutate(
    Age_Group = factor(Age_Group, levels = c("Young", "Older")),
    Response_group = factor(Response_group, levels = c("Poor responders", "Good responders"))  
  )

ggplot(plot_b_data, aes(y = Age_Group, x = percent, fill = Response_group)) +
  geom_col(position = "fill", colour = "black", width = 0.6) +
  scale_x_continuous(labels = scales::percent_format()) +
  scale_fill_manual(
    values = c("Poor responders"="#0072B2", "Good responders" = "#E69F00"),
    labels = c("PR", "GR")
  ) +
  labs(x = "Percent response", y = NULL, fill = NULL) +
  theme_bw() +
  theme(panel.grid = element_blank())
```

```{r}
#| message: false

ggsave(filename = "06_B2_plot.png",
       path = here("results"))
```

Plot C:

```{r}
plot_c_data <- analysis_data |>
  mutate(pre_GMT = exp(rowMeans(log(cbind(H1N1v1, Bv1, H3N2v1)))),
         log_pre_GMT = log(pre_GMT))

ggplot(plot_c_data, aes(x = Age_Group, y = log_pre_GMT)) +
  geom_boxplot(fill = "white", color = "black") +
  labs(x = NULL, y = "log(pre-GMT)") +
  theme_bw() +
  theme(panel.grid = element_blank(),
        legend.position = "none")
```

The geometric mean titer pre vaccine is higher in the young age group than in the older

```{r}
#| message: false

ggsave(filename = "06_C_plot.png",
       path = here("results"))
```
